pyproject.toml
[project]
name = "módulo_2_desenvolvimento_de_apis_em_python_com_flask"
version = "0.1.0"
description = ""
authors = [
    {name = "filipeCoutinho",email = "filipeccoutinho2017@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "flask-migrate",
    "jinja2 (==3.0.3)",
    "itsdangerous (==2.1.2)",
    "flask",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-sqlalchemy",
    "flask-bcrypt",
    "pytest"
]

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"